requirements.txt
python-telegram-bot
supabase
python-dotenv
google-generativeai
Pillow
Flask
